Create initial models, urls, views, templates and migrations
A pretty large commit but the resulting app is pretty straightforward.
The only 'visible' result is an unfinished welcome page and a lobby
which responds correctly if the person is logged in. All the other work
is provisional backend stuff.

There are some placeholder methods and attributes on the model which I
anticipate I'll need to finish later, although I still haven't figured
out how exactly the client-server communication is going to work (likely
via websockets) + I'm a little rusty at Django so I may have confused
some views and methods.

It seems very likely I'll need a few extra models for a working app, but
I'm not sure what those would be exactly so for now the essential
remains the only model.