Add basic log in/out w/ unfinished templates + no registration
Took forever to commit this because I was tearing my hair out over
nothing. Including the urls as I've done, and as was suggested by the
docs, overwrites the admin site templates except for login. Figuring out
why this is took a while, and while I now know at a technical level why
this happens I wasted a lot more time figuring out *why* this happens. I
still don't know that but have the good sense to just drop it.

The secondary templates are mainly copy-pastes because I'll need to go
back later and add a registration page anyway. I've checked and these
should exactly meet the requirements of the standard authentication
views.

The index view (currently wired to the web root) works nicely now,
asking for a login if needed before forwarding to the lobby. I've done
it the 'raw' way for now because it's pretty easy but later it might
make sense to have the lobby be the root of the app and use the
@login_required decorator instead.