Add lobby and game creation
Implementing a working lobby wasn't too bad, but the stuff to create
games was a bit of a nightmare. It's late now and I've earned going to
sleep.

The lobby is a straightforward view + template which doesn't need much
comment. Also added views and templates to: view a game's info prior to
joining; and create a game (new ModelForm required here too).

The urls were updated to include the new views and all views which
require it had the @login_required decorator added. It seems to work
like a charm!

Tweaking the model to add nulls and blanks was needed for two fields,
although I only realised blank was needed after migrating, hence the
duplicate migrations.