Add player joining functionality
Now I'm rocking and rolling. It's only taken an hour or two for this big
commit!

First change to implement this is a tweak to the model to remove the
numeric active players attribute (references to it have been updated,
hopefully everywhere) and allow many-to-many linking of users (players)
to the games.

The game templates have been updated to reflect the number of
players and a given player's options (already joined, able to join,
unable to join).

View added to join a player and the format of the game-specific ones was
tweaked. Not entirely happy with it still so may tweak later. Also
updated the view game view to collect dynamic player info.